+++ src/heat/heat_hpc.cu
@@ -9,18 +9,19 @@
 #include <cusp/print.h>
 #include "json/json.h"
 
-#include "file/nc_utilities.h"
+#include "dg/file/nc_utilities.h"
 #include "dg/algorithm.h"
-#include "geometries/geometries.h"
+#include "dg/geometries/geometries.h"
 
 #include "parameters.h"
 #include "heat.cuh"
 
 int main( int argc, char* argv[])
 {
-    ////Parameter initialisation ////////////////////////////////////////////
+    ////Parameter initialisation ///////////////////////////////////////
     Json::Value js, gs;
     Json::CharReaderBuilder parser;
+    //read input without comments
     parser["collectComments"] = false;
     std::string errs;
     if(!(( argc == 4) || ( argc == 5)) )
@@ -29,25 +30,27 @@ int main( int argc, char* argv[])
         std::cerr << "OR "<< argv[0]<<" [inputfile] [geomfile] [output.nc] \n";
         return -1;
     }
-    else 
+    else
     {
         std::ifstream is(argv[1]);
         std::ifstream ks(argv[2]);
-        parseFromStream( parser, is, &js, &errs); //read input without comments
-        parseFromStream( parser, ks, &gs, &errs); //read input without comments
+        parseFromStream( parser, is, &js, &errs);
+        parseFromStream( parser, ks, &gs, &errs);
     }
     const heat::Parameters p( js); p.display( std::cout);
     const dg::geo::solovev::Parameters gp(gs); gp.display( std::cout);
-    ////////////////////////////////set up computations///////////////////////////
+    ////////////////////////////set up computations//////////////////////
 
     double Rmin=gp.R_0-p.boxscaleRm*gp.a;
     double Zmin=-p.boxscaleZm*gp.a*gp.elongation;
-    double Rmax=gp.R_0+p.boxscaleRp*gp.a; 
+    double Rmax=gp.R_0+p.boxscaleRp*gp.a;
     double Zmax=p.boxscaleZp*gp.a*gp.elongation;
 
     //Make grids
-    dg::CylindricalGrid3d grid( Rmin,Rmax, Zmin,Zmax, 0, 2.*M_PI, p.n, p.Nx, p.Ny, p.Nz, p.bc, p.bc, dg::PER);  
-    dg::CylindricalGrid3d grid_out( Rmin,Rmax, Zmin,Zmax, 0, 2.*M_PI, p.n_out, p.Nx_out, p.Ny_out,p.Nz_out,p.bc, p.bc, dg::PER); 
+    dg::CylindricalGrid3d grid( Rmin,Rmax, Zmin,Zmax, 0, 2.*M_PI,
+        p.n, p.Nx, p.Ny, p.Nz, p.bcx, p.bcy, dg::PER);
+    dg::CylindricalGrid3d grid_out( Rmin,Rmax, Zmin,Zmax, 0, 2.*M_PI,
+        p.n_out, p.Nx_out, p.Ny_out,p.Nz_out,p.bcx, p.bcy, dg::PER);
     dg::DVec w3d =  dg::create::volume(grid);
     dg::DVec w3dout =  dg::create::volume(grid_out);
 
@@ -66,7 +69,7 @@ int main( int argc, char* argv[])
         file::NC_Error_Handle errin;
         int ncidin;
         errin = nc_open( argv[4], NC_NOWRITE, &ncidin);
-        ///////////////////read in and show inputfile und geomfile//////////////////
+        //////////////read in and show inputfile und geomfile////////////
         size_t length;
         errin = nc_inq_attlen( ncidin, NC_GLOBAL, "inputfile", &length);
         std::string inputin( length, 'x');
@@ -83,98 +86,70 @@ int main( int argc, char* argv[])
         parseFromStream( parser, is, &gs, &errs);
         const heat::Parameters pin(js);
         const dg::geo::solovev::Parameters gpin(gs);
-        double Rminin = gpin.R_0 - pin.boxscaleRm*gpin.a;
-        double Zminin =-pin.boxscaleZm*gpin.a*gpin.elongation;
-        double Rmaxin = gpin.R_0 + pin.boxscaleRp*gpin.a; 
-        double Zmaxin = pin.boxscaleZp*gpin.a*gpin.elongation;
-        dg::CylindricalGrid3d grid_in( Rminin,Rmaxin, Zminin,Zmaxin, 0, 2.*M_PI, pin.n, pin.Nx, pin.Ny, pin.Nz, pin.bc, pin.bc, dg::PER);
         size_t start3din[4]  = {pin.maxout, 0, 0, 0};
-        size_t count3din[4]  = {1, grid_in.Nz(), grid_in.n()*grid_in.Ny(), grid_in.n()*grid_in.Nx()};
-        std::string namesin[1] = {"T"}; 
-        int dataIDsin[1]; 
-        errin = nc_inq_varid(ncidin, namesin[0].data(), &dataIDsin[0]);      
-        errin = nc_get_vara_double( ncidin, dataIDsin[0], start3din, count3din, transferH.data());
-        Tend=(dg::DVec)transferH;
-        errin = nc_close(ncidin);     
+        size_t count3din[4]  = {1, pin.Nz, pin.n*pin.Ny, pin.n*pin.Nx};
+        std::string namesin = {"T"};
+        int dataIDin;
+        errin = nc_inq_varid(ncidin, namesin.data(), &dataIDin);
+        errin = nc_get_vara_double( ncidin, dataIDin, start3din,
+                                    count3din, transferH.data());
+        dg::assign(transferH, Tend);
+        errin = nc_close(ncidin);
     }
-    // /////////////////////create RHS 
+    // /////////////////////create RHS
     std::cout << "Constructing Feltor...\n";
-    heat::Explicit<dg::CylindricalGrid3d, dg::IDMatrix, dg::DMatrix, dg::DVec> exp( grid, p,gp); //initialize before diffusion!
+    heat::Explicit<dg::CylindricalGrid3d, dg::IDMatrix, dg::DMatrix, dg::DVec> ex( grid, p,gp); //initialize before diffusion!
     std::cout << "initialize implicit" << std::endl;
     heat::Implicit<dg::CylindricalGrid3d, dg::IDMatrix, dg::DMatrix, dg::DVec > diffusion( grid, p,gp);
     std::cout << "Done!\n";
 
-    /////////////////////The initial field///////////////////////////////////////////
+    //////////////////The initial field/////////////////////////////////
     //initial perturbation
     dg::Gaussian3d init0(gp.R_0+p.posX*gp.a, p.posY*gp.a, M_PI, p.sigma, p.sigma, p.sigma_z, p.amp);
-//      dg::Gaussian init0( gp.R_0+p.posX*gp.a, p.posY*gp.a, p.sigma, p.sigma, p.amp);
-//     dg::BathRZ init0(16,16,p.Nz,Rmin,Zmin, 30.,5.,p.amp);
-//     solovev::ZonalFlow init0(p, gp);
-//     dg::CONSTANT init0( 0.);
 
     //background profile
-    dg::geo::Nprofile prof(p.bgprofamp, p.nprofileamp, gp, dg::geo::solovev::Psip(gp)); //initial background profile
-    std::vector<dg::DVec> y0(1, dg::evaluate( prof, grid)), y1(y0); 
+    dg::geo::Nprofile prof(0, p.nprofileamp, gp, dg::geo::solovev::Psip(gp)); //initial background profile
+    dg::DVec y0(dg::evaluate( prof, grid)), y1(y0);
     //field aligning
-//     dg::CONSTANT gaussianZ( 1.);
     dg::GaussianZ gaussianZ( 0., p.sigma_z*M_PI, 1);
-    y1[0] = exp.ds().fieldaligned().evaluate( init0, gaussianZ, (unsigned)p.Nz/2, 3); //rounds =2 ->2*2-1
-//     y1[2] = dg::evaluate( gaussianZ, grid);
-//     dg::blas1::pointwiseDot( y1[1], y1[2], y1[1]);
-    //no field aligning
-//     y1[0] = dg::evaluate( init0, grid);
-    
-    dg::blas1::axpby( 1., y1[0], 1., y0[0]); //initialize ni
-    if (p.bc ==dg::DIR)    {
-            dg::blas1::transform(y0[0], y0[0], dg::PLUS<>(-1)); //initialize ni-1
-    }
-//     dg::blas1::pointwiseDot(diffusion.damping(),y0[0], y0[0]); //damp with gaussprofdamp
-    ///////////////////TIME STEPPER   
-    //RK solver
-//     dg::RK<4, std::vector<dg::DVec> >  rk( y0);
-    //SIRK solver
-    dg::Adaptive<dg::ARKStep<std::vector<dg::DVec>>> adaptive(y0, "ARK-4-2-3", grid.size(), p.eps_time);
-    double dt = 1e-5;
-    // dg::Karniadakis< std::vector<dg::DVec> > karniadakis( y0, y0[0].size(),1e-13);
-     //karniadakis.init( exp, diffusion, 0, y0, p.dt);
+    y1 = ex.ds().fieldaligned().evaluate( init0, gaussianZ, (unsigned)p.Nz/2, 3); //rounds =2 ->2*2-1
+    dg::blas1::axpby( 1., y1, 1., y0); //initialize y0
+    ///////////////////TIME STEPPER
+    dg::Adaptive<dg::ARKStep<dg::DVec>> adaptive(
+        y0, "ARK-4-2-3", grid.size(), p.eps_time);
+    double dt = p.dt, dt_new = dt;
+    // dg::Karniadakis< dg::DVec > karniadakis( y0, y0.size(),1e-13);
+     //karniadakis.init( ex, diffusion, 0, y0, p.dt);
 
-    exp.energies( y0);//now energies and potential are at time 0
-    dg::DVec T0 = dg::evaluate( dg::one, grid);  
-    dg::DVec T1 = dg::evaluate( dg::one, grid);  
-
-    dg::blas1::axpby( 1., y0[0], 0., T0); //initialize ni
-    double normT0 = dg::blas2::dot(  w3d, T0);
-    double error = 0.,relerror=0.;
+    ex.energies( y0);//now energies and potential are at time 0
+    dg::DVec T0 = y0, T1(T0);
+    double normT0 = dg::blas2::dot( T0, w3d, T0);
     /////////////////////////////set up netcdf for output/////////////////////////////////////
 
     file::NC_Error_Handle err;
     int ncid;
     err = nc_create( argv[3],NC_NETCDF4|NC_CLOBBER, &ncid);
-    std::string input = js.toStyledString(); 
+    std::string input = js.toStyledString();
     err = nc_put_att_text( ncid, NC_GLOBAL, "inputfile", input.size(), input.data());
-    std::string geom = gs.toStyledString(); 
+    std::string geom = gs.toStyledString();
     err = nc_put_att_text( ncid, NC_GLOBAL, "geomfile", geom.size(), geom.data());
     int dim_ids[4];
     err = file::define_dimensions( ncid, dim_ids, &tvarID, grid_out);
 
-    //field IDs
-    std::string names[1] = {"T"}; 
-    int dataIDs[1]; 
-    err = nc_def_var( ncid, names[0].data(), NC_DOUBLE, 4, dim_ids, &dataIDs[0]);
-
     //energy IDs
     int EtimeID, EtimevarID;
     err = file::define_time( ncid, "energy_time", &EtimeID, &EtimevarID);
-    int energyID, massID, energyIDs[1], dissID, dEdtID, accuracyID,errorID,relerrorID;
-    err = nc_def_var( ncid, "energy",   NC_DOUBLE, 1, &EtimeID, &energyID);
-    err = nc_def_var( ncid, "mass",   NC_DOUBLE, 1, &EtimeID, &massID);
-    std::string energies[1] = {"Se"}; 
-    err = nc_def_var( ncid, energies[0].data(), NC_DOUBLE, 1, &EtimeID, &energyIDs[0]);
-    err = nc_def_var( ncid, "dissipation",   NC_DOUBLE, 1, &EtimeID, &dissID);
-    err = nc_def_var( ncid, "dEdt",     NC_DOUBLE, 1, &EtimeID, &dEdtID);
-    err = nc_def_var( ncid, "accuracy", NC_DOUBLE, 1, &EtimeID, &accuracyID);
-    err = nc_def_var( ncid, "error", NC_DOUBLE, 1, &EtimeID, &errorID);
-    err = nc_def_var( ncid, "relerror", NC_DOUBLE, 1, &EtimeID, &relerrorID);
+    std::string names0d []= { "heat", "entropy", "dissipation",
+        "entropy_dissipation", "accuracy", "error", "relerror"};
+    std::map<std::string, int> id0d;
+    std::map<std::string, double> value0d;
+    for( auto name : names0d)
+        err = nc_def_var( ncid, name.data(), NC_DOUBLE, 1, &EtimeID, &id0d[name]);
+    std::string names3d [] = {"T"};
+    std::map<std::string, int> id3d;
+    std::map<std::string, dg::HVec> value3d;
+    for( auto name : names3d)
+        err = nc_def_var( ncid, name.data(), NC_DOUBLE, 4, dim_ids, &id3d[name]);
 
     err = nc_enddef(ncid);
     ///////////////////////////////////first output/////////////////////////
@@ -182,18 +157,13 @@ int main( int argc, char* argv[])
     size_t start[4] = {0, 0, 0, 0};
     size_t count[4] = {1, grid_out.Nz(), grid_out.n()*grid_out.Ny(), grid_out.n()*grid_out.Nx()};
 
-    //interpolate coarse 2 fine grid
-    dg::IDMatrix interpolatec2f = dg::create::interpolation( grid_out, grid); //c2f
-//     cusp::ell_matrix<int, double, cusp::device_memory> interpolatec2f = dg::create::ell_interpolation( grid_out, grid);
     //interpolate fine 2 coarse grid
     dg::IDMatrix interpolatef2c = dg::create::interpolation( grid, grid_out);//f2c
-//     cusp::ell_matrix<int, double, cusp::device_memory> interpolatef2c = dg::create::ell_interpolation( grid, grid_out); 
-    
-//     dg::blas2::symv( interpolatec2f, y0[0], transferD); //interpolate field
-    transferD =y0[0]; // dont interpolate field
+
+    transferD =y0; // dont interpolate field
     err = nc_open(argv[3], NC_WRITE, &ncid);
-    dg::blas1::transfer(transferD, transferH);
-    err = nc_put_vara_double( ncid, dataIDs[0], start, count, transferH.data());        
+    dg::assign(transferD, value3d["T"]);
+    err = nc_put_vara_double( ncid, id3d["T"], start, count, value3d["T"].data());
 
     double time = 0;
     err = nc_put_vara_double( ncid, tvarID, start, count, &time);
@@ -201,36 +171,27 @@ int main( int argc, char* argv[])
 
     size_t Estart[] = {0};
     size_t Ecount[] = {1};
-    double energy0 = exp.energy(), mass0 = exp.mass(), E0 = energy0, mass = mass0, E1 = 0.0, dEdt = 0., diss = 0., accuracy=0.;
-    dg::blas1::axpby( 1., y0[0], -1.,T0, T1);
-    error = sqrt(dg::blas2::dot( w3d, T1)/normT0);
-    double diss0 =exp.energy_diffusion(); 
-    
+    double entropy0 = ex.entropy(), heat0 = ex.energy(); //at time 0
+    double E0 = entropy0;
+    dg::blas1::axpby( 1., y0, -1.,T0, T1);
+
     //Compute error to reference solution
+    value0d["error"] = sqrt(dg::blas2::dot( w3d, T1)/normT0);
+    value0d["relerror"]=0.;
     if (argc==5)
     {
-//         interpolate fine grid one coarse grid
+        // interpolate fine grid one coarse grid
         dg::blas2::symv( interpolatef2c, Tend, Tendc);
-        dg::blas1::axpby( 1., y0[0], -1.,Tendc,transfer);
-        relerror = sqrt(dg::blas2::dot( w3d, transfer)/dg::blas2::dot(w3dout,Tend));      
-//         interpolate coarse on fine grid
-/*        dg::blas2::symv( interpolatec2f, y0[0], transferD);
-       dg::blas1::axpby( 1., transferD, -1.,Tend,transferD);
-        relerror = sqrt(dg::blas2::dot( w3dout, transferD)/dg::blas2::dot( w3dout, Tend));   */           
+        dg::blas1::axpby( 1., y0, -1.,Tendc,transfer);
+        value0d["relerror"] = sqrt(dg::blas2::dot( w3d, transfer)/dg::blas2::dot(w3dout,Tend));
     }
-    std::vector<double> evec = exp.energy_vector();
-    double Se0 = evec[0];
-    double senorm = evec[0]/Se0;
-    double dEdtnorm = dEdt/Se0;
-    double dissnorm = diss/diss0;
-    err = nc_put_vara_double( ncid, energyID, Estart, Ecount, &energy0);
-    err = nc_put_vara_double( ncid, massID,   Estart, Ecount, &mass0);
-    err = nc_put_vara_double( ncid, energyIDs[0], Estart, Ecount, &senorm);
-    err = nc_put_vara_double( ncid, dissID,     Estart, Ecount,&dissnorm);
-    err = nc_put_vara_double( ncid, dEdtID,     Estart, Ecount,&dEdtnorm);
-    err = nc_put_vara_double( ncid, accuracyID, Estart, Ecount,&accuracy);
-    err = nc_put_vara_double( ncid, errorID, Estart, Ecount,&error);
-    err = nc_put_vara_double( ncid, relerrorID, Estart, Ecount,&relerror);
+    value0d["heat"] = ex.energy();
+    value0d["entropy"] = ex.entropy();
+    value0d["dissipation"] = ex.energy_diffusion();
+    value0d["entropy_dissipation"] = ex.entropy_diffusion();
+    value0d["accuracy"] = 0.;
+    for( auto name : names0d)
+        err = nc_put_vara_double( ncid, id0d[name], Estart, Ecount, &value0d[name]);
     err = nc_close(ncid);
     std::cout << "First write successful!\n";
 
@@ -248,56 +209,54 @@ int main( int argc, char* argv[])
         for( unsigned j=0; j<p.itstp; j++)
         {
             try{
-//                 rk.step( exp, time,y0, time,y0, p.dt); //RK stepper
-                adaptive.step(exp,diffusion,time,y0,time,y0,dt, dg::pid_control, dg::l2norm, 1e-5, 1e-10); //SIRK stepper
-                 //karniadakis.step( exp, diffusion, time, y0);  //Karniadakis stepper
+//                 rk.step( ex, time,y0, time,y0, p.dt); //RK stepper
+                dt = dt_new;
+                adaptive.step(ex,diffusion,time,y0,time,y0,dt_new, dg::pid_control, dg::l2norm, 1e-5, 1e-10);
+                 //karniadakis.step( ex, diffusion, time, y0);  //Karniadakis stepper
               }
-              catch( dg::Fail& fail) { 
+              catch( dg::Fail& fail) {
                 std::cerr << "CG failed to converge to "<<fail.epsilon()<<"\n";
                 std::cerr << "Does Simulation respect CFL condition?\n";
                 err = nc_close(ncid);
                 return -1;}
             step++;
-            exp.energies(y0);//advance potential and energies
+            ex.energies(y0);//advance potential and energies
             Estart[0] = step;
-            E1 = exp.energy(), mass = exp.mass(), diss = exp.energy_diffusion();
-            dEdt = (E1 - E0)/p.dt; 
-            E0 = E1;
-            dg::blas1::axpby( 1., y0[0], -1.,T0, T1);
-            error = sqrt(dg::blas2::dot( w3d, T1)/normT0);
+            value0d["entropy"] = ex.entropy();
+            value0d["heat"] = ex.energy();
+            value0d["entropy_dissipation"] = ex.entropy_diffusion();
+            value0d["heat_dissipation"] = ex.energy_diffusion();
+            double dEdt = (value0d["entropy"] - E0)/dt;
+            value0d["accuracy"] = 2.*fabs(
+                            (dEdt - value0d["entropy_dissipation"])/
+                            (dEdt + value0d["entropy_dissipation"]));
+            E0 = value0d["entropy"];
+            //compute errors
+            dg::blas1::axpby( 1., y0, -1.,T0, T1);
+            value0d["error"] = sqrt(dg::blas2::dot( w3d, T1)/normT0);
             if (argc==5)
             {
-//         interpolate fine on coarse grid
+                //interpolate fine on coarse grid
                 dg::blas2::symv( interpolatef2c, Tend, Tendc);
-                dg::blas1::axpby( 1., y0[0], -1.,Tendc,transfer);
-                relerror = sqrt(dg::blas2::dot(w3d, transfer)/dg::blas2::dot(w3dout,Tend));  
-//         interpolate coarse on fine grid
-//                 dg::blas2::symv( interpolatec2f, y0[0], transferD);
-//                 dg::blas1::axpby( 1., transferD, -1.,Tend,transferD);
-//                 relerror = sqrt(dg::blas2::dot( w3dout, transferD)/dg::blas2::dot( w3dout, Tend)); 
+                dg::blas1::axpby( 1., y0, -1.,Tendc,transfer);
+                value0d["relerror"] = sqrt(dg::blas2::dot(w3d, transfer)/dg::blas2::dot(w3dout,Tend));
 
             }
-            accuracy = 2.*fabs( (dEdt-diss)/(dEdt + diss));
-            evec = exp.energy_vector();
-            senorm = evec[0]/Se0;
-            dEdtnorm = dEdt/Se0;
-            dissnorm = diss/diss0;
             err = nc_open(argv[3], NC_WRITE, &ncid);
             err = nc_put_vara_double( ncid, EtimevarID, Estart, Ecount, &time);
-            err = nc_put_vara_double( ncid, energyID, Estart, Ecount, &E1);
-            err = nc_put_vara_double( ncid, massID,   Estart, Ecount, &mass);
-            err = nc_put_vara_double( ncid, energyIDs[0], Estart, Ecount, &senorm);
-            err = nc_put_vara_double( ncid, dissID,     Estart, Ecount,&dissnorm);
-            err = nc_put_vara_double( ncid, dEdtID,     Estart, Ecount,&dEdtnorm);
-            err = nc_put_vara_double( ncid, accuracyID, Estart, Ecount,&accuracy);
-            err = nc_put_vara_double( ncid, errorID, Estart, Ecount,&error);
-            err = nc_put_vara_double( ncid, relerrorID, Estart, Ecount,&relerror);
+            for( auto name : names0d)
+                err = nc_put_vara_double( ncid, id0d[name], Estart, Ecount, &value0d[name]);
 
-            std::cout << "(m_tot-m_0)/m_0: "<< (exp.mass()-mass0)/mass0<<"\t";
-            std::cout << "(E_tot-E_0)/E_0: "<< (E1-energy0)/energy0<<"\t";
-            std::cout <<" d E/dt = " << dEdt <<" Lambda = " << diss << " -> Accuracy: "<< accuracy << " -> error2t0: "<< error <<" -> error2ref: "<< relerror <<"\n";
+            std::cout <<"(Q_tot-Q_0)/Q_0: "
+                      << (ex.energy()-heat0)/heat0<<"\t";
+            std::cout <<"(E_tot-E_0)/E_0: "
+                      << (ex.entropy()-entropy0)/entropy0<<"\t";
+            std::cout <<" d E/dt = " << dEdt
+                      <<" Lambda = " << ex.entropy_diffusion()
+                      <<" -> Accuracy: "<< value0d["accuracy"]
+                      <<" -> error2t0: "<< value0d["error"]
+                      <<" -> error2ref: "<< value0d["relerror"] <<"\n";
             err = nc_close(ncid);
-
         }
 #ifdef DG_BENCHMARK
         ti.toc();
@@ -307,15 +266,14 @@ int main( int argc, char* argv[])
         //////////////////////////write fields////////////////////////
         start[0] = i;
 
-//         dg::blas2::symv( interpolatec2f, y0[0], transferD); //interpolate field
-        transferD =y0[0]; //dont interpolate field
-        dg::blas1::transfer(transferD, transferH);
+        transferD = y0; //dont interpolate field
+        dg::assign(transferD, value3d["T"]);
         err = nc_open(argv[3], NC_WRITE, &ncid);
-        err = nc_put_vara_double( ncid, dataIDs[0], start, count, transferH.data());        
+        err = nc_put_vara_double( ncid, id3d["T"], start, count, value3d["T"].data());
         err = nc_put_vara_double( ncid, tvarID, start, count, &time);
         err = nc_close(ncid);
     }
-    t.toc(); 
+    t.toc();
     unsigned hour = (unsigned)floor(t.diff()/3600);
     unsigned minute = (unsigned)floor( (t.diff() - hour*3600)/60);
     double second = t.diff() - hour*3600 - minute*60;